+++ dapt/db/__init__.py
@@ -2,3 +2,5 @@
 from .database import Database
 from .delimited_file import Delimited_file
 from .sheets import Sheet
+
+