+++ tests/test_box.py
@@ -1,0 +1,30 @@
+"""
+Test the Box class in `dapt.storage.box`
+"""
+
+import dapt
+
+# Test if a file can be uploaded
+def test_box_upload_file():
+    pass
+
+# Test if a folder can be uploaded
+def test_box_upload_folder():
+    pass
+
+# Test if the uploaded file can be dowloaded and is the same as the original
+def test_box_download_file():
+    pass
+
+# Test if the uploaded folder can be dowloaded and is the same as the original
+def test_box_download_folder():
+    pass
+
+# Test if the file can be deleted
+def test_box_delete_file():
+    pass
+
+# Test if the folder can be deleted
+def test_box_delete_folder():
+    pass
+